Out-of-source builds without environment variables.
Using CMake's configure_file command to generate a header file defining
a C string containing the path to the source root directory; the
directory where the resource folder is. Textures and models can now be
found directly as long the user doesn't move stuff around.

Shaders are still a problem.