Consider using smart pointers instead of deleting memory with API
Basically, prerequisite of this project is C++17 so we might want to use some modern way of memory handling here. 

Using raw pointers is considered not the best practice at all cause it increases the risk of memory-managment related problems. 

In this particular case it makes code a bit easier as well